requirements.txt
streamlit
openai
pypdf
python-docx
python-pptx
pandas
openpyxl
fpdf2